@@ -1,4 +1,3 @@
-from __future__ import absolute_import
 # Copyright (c) 2010-2019 openpyxl
 
 """Manage individual cells in a spreadsheet.
@@ -19,9 +18,6 @@ import re
 from itertools import islice, product
 
 from openpyxl.compat import (
-    unicode,
-    basestring,
-    bytes,
     NUMERIC_TYPES,
     deprecated,
 )
@@ -68,7 +64,7 @@ try:
 except ImportError:
     pass
 
-STRING_TYPES = (basestring, unicode, bytes)
+STRING_TYPES = (str, bytes)
 KNOWN_TYPES = NUMERIC_TYPES + TIME_TYPES + STRING_TYPES + (bool, type(None))
 
 ILLEGAL_CHARACTERS_RE = re.compile(r'[\000-\010]|[\013-\014]|[\016-\037]')
@@ -158,10 +154,10 @@ class Cell(StyleableObject):
         """Check string coding, length, and line break character"""
         if value is None:
             return
-        # convert to unicode string
-        if not isinstance(value, unicode):
-            value = unicode(value, self.encoding)
-        value = unicode(value)
+        # convert to str string
+        if not isinstance(value, str):
+            value = str(value, self.encoding)
+        value = str(value)
         # string must never be longer than 32,767 characters
         # truncate if necessary
         value = value[:32767]
@@ -172,7 +168,7 @@ class Cell(StyleableObject):
     def check_error(self, value):
         """Tries to convert Error" else N/A"""
         try:
-            return unicode(value)
+            return str(value)
         except UnicodeDecodeError:
             return u'#N/A'
 
@@ -222,7 +218,7 @@ class Cell(StyleableObject):
 
     def _infer_value(self, value):
         """Given a string, infer type and formatting options."""
-        if not isinstance(value, unicode):
+        if not isinstance(value, str):
             value = str(value)
 
         # number detection
