@@ -1,7 +1,5 @@
-from __future__ import absolute_import
 # Copyright (c) 2010-2019 openpyxl
 
-from openpyxl.compat import unicode
 
 from openpyxl.descriptors.serialisable import Serialisable
 from openpyxl.descriptors import (
@@ -14,7 +12,7 @@ class AuthorList(Serialisable):
 
     tagname = "authors"
 
-    author = Sequence(expected_type=unicode)
+    author = Sequence(expected_type=str)
     authors = Alias("author")
 
     def __init__(self,
