@@ -1,5 +1,5 @@
-from __future__ import absolute_import
 # Copyright (c) 2010-2019 openpyxl
+
 
 from openpyxl.workbook import Workbook
 from openpyxl.tests.helper import compare_xml
