@@ -1,5 +1,4 @@
-from __future__ import absolute_import
-# Copyright (c) 2010-2019 openpyxl
+#copyright openpyxl 2010-2015
 
 """
 Generic serialisable classes
