@@ -1,10 +1,9 @@
-from __future__ import absolute_import
 # Copyright (c) 2010-2019 openpyxl
 
 import pytest
 
 # compatibility imports
-from openpyxl.compat import unicode
+
 
 # package imports
 from openpyxl.styles import numbers
