@@ -1,6 +1,4 @@
-from __future__ import absolute_import
 # Copyright (c) 2010-2019 openpyxl
-
 
 """Write the theme xml based on a fixed string."""
 
@@ -287,3 +285,7 @@ theme_xml = """<?xml version="1.0"?>
   <a:extraClrSchemeLst/>
 </a:theme>
 """
+
+def write_theme():
+    """Write the theme xml."""
+    return theme_xml
